requirements.txt
streamlit
pandas
numpy
matplotlib
langchain
faiss-cpu
PyPDF2
openai
